Added classes ScreenSetBuilder and ScreenBuilder to lighten class Event
